backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
groq>=0.30.0,<1.0.0
python-dotenv==1.1.1